[#162] refactor: move Tonic extensions for chord handling
